requirements.txt
streamlit>=1.32.0
boto3>=1.34.0
botocore>=1.34.0
langchain>=0.3.1
langchain-community>=0.0.33
langchain-core>=0.1.32
langchain-ollama>=0.0.4
ollama>=0.1.6
plotly>=6.0.1
requests>=2.31.0
azure-ai-projects
azure-identity
matplotlib>=3.7.2
pandas>=2.0.0
openpyxl>=3.1.2
